+++ kit/entry/server_prod.js
@@ -27,11 +27,8 @@ import server, { createReactHandler, staticMiddleware } from './server';
 // ----------------------
 
 // Read in manifest files
-const [manifest, chunkManifest] = ['manifest', 'chunk-manifest'].map(
-  name => JSON.parse(
-    readFileSync(path.resolve(PATHS.dist, `${name}.json`), 'utf8'),
-  ),
-);
+const [manifest, chunkManifest] = ['manifest', 'chunk-manifest']
+  .map(name => JSON.parse(readFileSync(path.resolve(PATHS.dist, `${name}.json`), 'utf8')));
 
 // Get manifest values
 const css = manifest['browser.css'];