+++ kit/views/ssr.js
@@ -9,24 +9,29 @@ import PropTypes from 'prop-types';
 
 // ----------------------
 
-const Html = ({ head, scripts, window, css, children }) => (
-  <html lang="en" prefix="og: http://ogp.me/ns#">
+const Html = ({ helmet, scripts, window, css, children }) => (
+  <html lang="en" prefix="og: http://ogp.me/ns#" {...helmet.htmlAttributes.toString()}>
     <head>
+      {helmet.title.toComponent()}
       <meta charSet="utf-8" />
       <meta httpEquiv="X-UA-Compatible" content="IE=edge" />
       <meta httpEquiv="Content-Language" content="en" />
       <meta name="viewport" content="width=device-width, initial-scale=1" />
-      {head.meta.toComponent()}
+      {helmet.meta.toComponent()}
       <link rel="stylesheet" href={css} />
-      {head.title.toComponent()}
+      {helmet.link.toComponent()}
+      {helmet.style.toComponent()}
+      {helmet.script.toComponent()}
+      {helmet.noscript.toComponent()}
     </head>
-    <body>
+    <body {...helmet.bodyAttributes.toComponent()}>
       <div id="main">{children}</div>
       <script
         dangerouslySetInnerHTML={{
           __html: Object.keys(window).reduce(
             (out, key) => out += `window.${key}=${JSON.stringify(window[key])};`,
-            ''),
+            '',
+          ),
         }} />
       {scripts.map(src => <script key={src} src={src} />)}
     </body>
@@ -34,7 +39,7 @@ const Html = ({ head, scripts, window, css, children }) => (
 );
 
 Html.propTypes = {
-  head: PropTypes.object.isRequired,
+  helmet: PropTypes.object.isRequired,
   window: PropTypes.object.isRequired,
   scripts: PropTypes.arrayOf(PropTypes.string).isRequired,
   css: PropTypes.string.isRequired,