+++ src/app.js
@@ -77,8 +77,8 @@ if (SERVER) {
     Uncomment the next two lines to enable SSL!
   */
 
-  const cert = require('src/cert/self_signed');
-  config.enableSSL({ key: cert.key, cert: cert.cert });
+  // const cert = require('src/cert/self_signed');
+  // config.enableSSL({ key: cert.key, cert: cert.cert });
 
   // If wanted, you could also run an *SSL-only* server by uncommenting:
   // config.disableHTTP();